add useTypingTask hook
